% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdpla-package.R
\docType{data}
\name{country_codes}
\alias{country_codes}
\title{Country codes}
\description{
A data.frame (249 rows, 2 columns) containing all language codes.
The columns are as follows:
}
\details{
\itemize{
  \item code. country code
  \item name. country name
}
}
\keyword{datasets}

