% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dpla_key}
\alias{dpla_key}
\title{Request an API key from the Digital Public Library of America (DPLA).}
\usage{
dpla_key(email, ...)
}
\arguments{
\item{email}{(character) An email address.}

\item{...}{Curl options passed on to \code{\link[httr]{POST}}, eg., \code{\link[httr]{verbose}}}
}
\value{
On success, a message that your API key will be emailed to you.
}
\description{
Request an API key from the Digital Public Library of America (DPLA).
}
\examples{
\donttest{
dpla_key(email="stuff@thing.com")
dpla_key(email="scott.chamberlain@berkeley.edu")
}
}

