% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdpla-package.R
\docType{data}
\name{language_codes}
\alias{language_codes}
\title{Language codes.}
\description{
A data.frame (7879 rows, 2 columns) containing all language codes.
The columns are as follows:
}
\details{
\itemize{
  \item id. id (or code) of the language
  \item name. longer name of the language
}
}
\keyword{datasets}

