% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdpla-package.R
\docType{package}
\name{rdpla-package}
\alias{rdpla-package}
\alias{rdpla}
\title{R Client for the Digital Public Library of America (DPLA).}
\description{
Interact with the Digital Public Library of America (DPLA)
REST API from R, including search.

For an introduction to \pkg{rdpla}, see the vignette
\strong{Introduction to rdpla}
}
\section{Package API}{


The following are the main functions in \pkg{rdpla}

\itemize{
\item dpla_items - Work with the DPLA items endpoint
\item dpla_collections - Work with the DPLA collections endpoint
\item dpla_bulk - Download bulk and compressed JSON data
}
}

\section{Authentication}{

See \code{\link[=dpla_get_key]{dpla_get_key()}} for authentication help.
}

\author{
Scott Chamberlain
}
