% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.r
\name{collections}
\alias{collections}
\title{Search collections from the Digital Public Library of America (DPLA).}
\usage{
collections(q = NULL, title = NULL, description = NULL, fields = NULL,
  sort_by = NULL, page_size = 10, page = NULL, key = NULL, ...)
}
\arguments{
\item{q}{Query terms.}

\item{title}{Query in the title field}

\item{description}{Query in the description field}

\item{fields}{A vector of the fields to return in the output. The default}

\item{sort_by}{The default sort order is ascending. Most, but not all fields
can be sorted on. Attempts to sort on an un-sortable field will return
the standard error structure with a HTTP 400 status code.
is all fields. See details for options.}

\item{page_size}{Number of items to return, defaults to 10. Max of 500.}

\item{page}{Page number to return, defaults to NULL.}

\item{key}{Your DPLA API key. Either pass in here, store as en env var in either
\code{.Renviron}/\code{.bash_profile}/etc., or in your \code{.Rprofile} file
and it will be read in on function execution.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\value{
A list with a slot for metadata (meta) and data (data).
}
\description{
Search collections from the Digital Public Library of America (DPLA).
}
\examples{
\donttest{
collections(q="university")
collections(q="university of texas", page_size=2)
collections(q="university of texas", fields='id', page_size=2)
collections(q="university of texas", sort_by='title', page_size=5)
collections(title="paso")
collections(description="east")
}
}

