% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dpla_collections}
\alias{dpla_collections}
\title{Search collections from the Digital Public Library of America (DPLA).}
\usage{
dpla_collections(q = NULL, fields = NULL, sort_by = NULL, limit = 10,
  page = NULL, key = getOption("dplakey"), ...)
}
\arguments{
\item{q}{Query terms.}

\item{fields}{A vector of the fields to return in the output. The default}

\item{sort_by}{The default sort order is ascending. Most, but not all fields
can be sorted on. Attempts to sort on an un-sortable field will return
the standard error structure with a HTTP 400 status code.
is all fields. See details for options.}

\item{limit}{Number of items to return, defaults to 10. Max of 100.}

\item{page}{Page number to return, defaults to NULL.}

\item{key}{Your DPLA API key. Either pass in here, or store in your \code{.Rprofile} file
and it will be read in on function execution.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\value{
xxxx
}
\description{
Search collections from the Digital Public Library of America (DPLA).
}
\examples{
\donttest{
dpla_collections(q="university")
dpla_collections(q="university of texas", limit=2)
dpla_collections(q="university of texas", fields='id', limit=2)
dpla_collections(q="university of texas", sort_by='title', limit=5)
}
}

