% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dpla_items}
\alias{dpla_items}
\title{Search items from the Digital Public Library of America (DPLA).}
\usage{
dpla_items(q = NULL, verbose = FALSE, fields = NULL, limit = 10,
  page = NULL, sort_by = NULL, date.before = NULL, date.after = NULL,
  key = getOption("dplakey"), ...)
}
\arguments{
\item{q}{Query terms.}

\item{verbose}{If TRUE, fun little messages print to console to inform you
of things.}

\item{fields}{A vector of the fields to return in the output. The default
is all fields. See details for options.}

\item{limit}{Number of items to return, defaults to 10. Max of 100.}

\item{page}{Page number to return, defaults to NULL.}

\item{sort_by}{The default sort order is ascending. Most, but not all fields
can be sorted on. Attempts to sort on an un-sortable field will return
the standard error structure with a HTTP 400 status code.}

\item{date.before}{Date before}

\item{date.after}{Date after}

\item{key}{Your DPLA API key. Either pass in here, or store in your \code{.Rprofile} file
and it will be read in on function execution.}

\item{...}{Curl options passed on to \code{\link[httr]{GET}}}
}
\value{
A list of length two: meta with the metadata for the call (found, offset [aka start],
and limit [number results returned]), and the resulting data.frame of results.
}
\description{
Search items from the Digital Public Library of America (DPLA).
}
\details{
Options for the fields argument are:
\itemize{
 \item id The item id
 \item title The title of the object
 \item decription The description of the object
 \item subject The subjects of the object
 \item creator The creator of the object
 \item type The type of the object
 \item publisher The publisher of the object
 \item format The format of the object
 \item rights The rights for the object
 \item contributor The contributor of the object
 \item spatial The spatial of the object
 \item isPartOf The isPartOf thing, not sure what this is
 \item provider The provider of the object
}
}
\examples{
\donttest{
# Basic search, "fruit" in any fields
dpla_items(q="fruit")

# Limit records returned
dpla_items(q="fruit", limit=2)

# Some verbosity
dpla_items(q="fruit", verbose=TRUE, limit=2)

# Return certain fields
dpla_items(q="fruit", verbose=TRUE, fields=c("id","publisher","format"))
dpla_items(q="fruit", fields="subject")

# Max is 100 per call, but the function handles larger numbers by looping
dpla_items(q="fruit", fields="id", limit=200)
dpla_items(q="fruit", fields=c("id","provider"), limit=200)
out <- dpla_items(q="science", fields=c("id","subject"), limit=400)
head(out$data)

# Search by date
out <- dpla_items(q="science", date.before=1900, limit=200)
head(out$data)

# Spatial search
dpla_items(q='Boston', fields='spatial')
}
}

