% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{rdpla-package}
\alias{rdpla-package}
\title{R Client for the Digital Public Library of America (DPLA).}
\description{
Interact with the Digital Public Library of America (DPLA) REST API from R,
including search, etc, etc.
}
\details{
\tabular{ll}{
 Package: \tab rdpla\cr
 Type: \tab Package\cr
 Version: \tab 0.0.1.99\cr
 Date: \tab 2014-10-28\cr
 License: \tab MIT\cr
}
}
\examples{
\donttest{
print("foo")
}
}
\author{
Scott Chamberlain
}
\seealso{
\code{\link[rmetadata]{rmetadata}} \code{\link[reuropeana]{reuropeana}}
\code{\link[rhathi]{rhathi}}
}

