% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdpla-package.R
\docType{package}
\name{rdpla-package}
\alias{rdpla}
\alias{rdpla-package}
\title{R Client for the Digital Public Library of America (DPLA).}
\description{
Interact with the Digital Public Library of America (DPLA)
REST API from R, including search.

For an introduction to \pkg{rdpla}, see the vignette
\strong{Introduction to rdpla}
}
\section{Authentication}{

See \code{\link{dpla_get_key}} for authentication help.
}
\author{
Scott Chamberlain
}

