% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdpla-package.R
\docType{package}
\name{rdpla-package}
\alias{rdpla}
\alias{rdpla-package}
\title{R Client for the Digital Public Library of America (DPLA).}
\description{
Interact with the Digital Public Library of America (DPLA) REST API from R,
including search, etc, etc.
}
\details{
\tabular{ll}{
 Package: \tab rdpla\cr
 Type: \tab Package\cr
 Version: \tab 0.0.1.99\cr
 Date: \tab 2014-10-28\cr
 License: \tab MIT\cr
}
}
\examples{
\donttest{
print("foo")
}
}
\author{
Scott Chamberlain
}

